fix(#1117): extract Auth0ProviderWithNavigate to its own module
Lint error react-refresh/only-export-components: a component defined
in main.tsx (which also runs createRoot) breaks Vite fast-refresh.
Move the wrapper to frontend/src/Auth0ProviderWithNavigate.tsx so
main.tsx is import-side-effects only.
